% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small-helpers.R
\name{get_chr}
\alias{get_chr}
\title{Retrieve a value from a Python \code{dict} as a character}
\usage{
get_chr(obj, key)
}
\arguments{
\item{obj}{Python \code{dict} or object with a \code{get()} method}

\item{key}{the key to retrive the value for}
}
\description{
It can be cumbersome to deal with all of the \code{$} in inane Python code.
This function lets you specify the object (a Python \code{dict} or a Python object
with a \code{get()} method) and the key and this will return the value or
\code{NA_character_} if there was no key..
}
