% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small-helpers.R
\name{psytr}
\alias{psytr}
\title{Compactly Display the Structure of a Python object}
\usage{
psytr(obj, ...)
}
\arguments{
\item{obj}{the python object}

\item{...}{passed to \code{\link[utils:str]{utils::str()}}}
}
\description{
Runs \code{\link[utils:str]{utils::str()}} on the object dictionary
}
\examples{
\dontrun{
library(reticulate)

mailbox = import("mailbox")
pystr(mailbox)
}
}
