% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small-helpers.R
\name{pynames}
\alias{pynames}
\title{(All) The Names of a Python Object}
\usage{
pynames(obj)
}
\arguments{
\item{obj}{the python object}
}
\description{
Running [names()] on a Python object provides the public interface names.
This function shows all the hidden names as well
}
\examples{
\dontrun{
library(reticulate)

mailbox = import("mailbox")
names(mailbox)
pynames(mailbox)
}
}
