% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-the-docs.R
\name{readthedocs}
\alias{readthedocs}
\title{Use "Read the Docs" to get help on a Python module/object}
\usage{
readthedocs(obj, py_ver = 3, check_keywords = TRUE)
}
\arguments{
\item{obj}{the unquoted Python object or a character string with the query keyword}

\item{check_keywords}{logical; if \code{TRUE} \code{check_keywords=yes} will be
used in the \code{docs.python.org} URL}

\item{pyver}{defautls to \code{3} but you can use any Python version string}
}
\description{
Uses \code{\link[utils:browseURL]{utils::browseURL()}} to open up a browser window pointing to
a \url{https://docs.python.org} search for the object in question.
}
\details{
You pass in the object itself (the function will handle the rest) or
a character string with the item to lookup.
}
\examples{
\dontrun{
library(reticulate)

mailbox = import("mailbox")
readthedocs(mailbox)

# or

readthedocs("mailbox")
}
}
