% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanipy-package.R
\docType{package}
\name{sanipy}
\alias{sanipy}
\alias{sanipy-package}
\title{Bring some sanity to R when working with Python via reticulate}
\description{
A set of utilities and helpers that make using Python via \code{reiculate} a bit saner.
}
\author{
Bob Rudis (bob@rud.is)
}
