% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small-helpers.R
\name{utf8_decode}
\alias{utf8_decode}
\title{Decode a Python structure and retrieve a UTF-8 encoded value from it}
\usage{
utf8_decode(obj)
}
\arguments{
\item{obj}{Python object}
}
\description{
Calls \code{obj$decode("utf-8", "ignore")} and returns the value or \code{NA_character_}
if the result is (Python) \code{None}.
}
