% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{zlength-zero-default}
\alias{zlength-zero-default}
\alias{\%|0|\%}
\title{Default value for an object length being 0}
\usage{
x \%|0|\% y
}
\arguments{
\item{x, y}{if \code{x} has length \code{0}, will return \code{y}; otherwise it returns \code{x}}
}
\description{
The \code{purrr} package has \code{\%||\%} which is super-useful when wanting to
infix a default value for an object being \code{NULL}. This function does
the same for \code{length(x) == 0}.
}
\examples{
1 \%|0|\% 2
NULL \%|0|\% 2
character(0) \%|0|\% 2
}
